Just Commit

+++ Simple Basic Problem solving/SmallFactorial.c
@@ -1,5 +1,5 @@
 #include <stdio.h>
-double factorial(int a);
+long long int factorial(int a);
 int main()
 {
   int T;
@@ -11,14 +11,14 @@ int main()
   }
   for(int i=0; i<T; i++)
   {
-    printf("%.0lf\n",factorial(arr[i]));
+    printf("%lld\n",factorial(arr[i]));
   }
 
   return 0;
 }
-double factorial(int a)
+long long int factorial(int a)
 {
-  double fact=1;
+  int fact=1;
   if(a==0 || a==1)
   {
     return 1;
